fixes_completed
исправил ошибки
